Run tests in a constant order
We have sporadic failures on some of the tests. I suspect that some
of the tests don't leave the cluster in a stable state (for example
node_eviction test).

In order to not block other PRs on CI, adding a temporary solution for
running the tests in a constant order.

Signed-off-by: gbenhaim <galbh2@gmail.com>